+++ libnymea-app/ruletemplates/calendaritemtemplate.h
@@ -66,7 +66,7 @@ class CalendarItemTemplates: public QAbstractListModel
 
 public:
     CalendarItemTemplates(QObject *parent = nullptr): QAbstractListModel(parent) {}
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return m_list.count(); }
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return static_cast<int>(m_list.count()); }
     QVariant data(const QModelIndex &index, int role) const override { Q_UNUSED(index); Q_UNUSED(role); return QVariant(); }
 
     Q_INVOKABLE CalendarItemTemplate* get(int index) const {
@@ -78,7 +78,7 @@ public:
 
     void addCalendarItemTemplate(CalendarItemTemplate *calendarItemTemplate) {
         calendarItemTemplate->setParent(this);
-        beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+        beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
         m_list.append(calendarItemTemplate);
         endInsertRows();
     }
