+++ libnymea-app/ruletemplates/eventdescriptortemplate.h
@@ -71,12 +71,12 @@ public:
     EventDescriptorTemplates(QObject *parent = nullptr): QAbstractListModel(parent) {}
 
     QStringList interfaces() const;
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return m_list.count(); }
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return static_cast<int>(m_list.count()); }
     QVariant data(const QModelIndex &index, int role) const override { Q_UNUSED(index); Q_UNUSED(role); return QVariant(); }
 
     void addEventDescriptorTemplate(EventDescriptorTemplate *eventDescriptorTemplate) {
         eventDescriptorTemplate->setParent(this);
-        beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+        beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
         m_list.append(eventDescriptorTemplate);
         endInsertRows();
         emit countChanged();
