+++ libnymea-app/ruletemplates/ruleactionparamtemplate.h
@@ -66,12 +66,12 @@ class RuleActionParamTemplates : public QAbstractListModel
 public:
     explicit RuleActionParamTemplates(QObject *parent = nullptr): QAbstractListModel(parent) {}
 
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return m_list.count(); }
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return static_cast<int>(m_list.count()); }
     QVariant data(const QModelIndex &index, int role) const override { Q_UNUSED(index) Q_UNUSED(role) return QVariant(); }
 
     void addRuleActionParamTemplate(RuleActionParamTemplate *ruleActionParamTemplate) {
         ruleActionParamTemplate->setParent(this);
-        beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+        beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
         m_list.append(ruleActionParamTemplate);
         endInsertRows();
         emit countChanged();
