+++ libnymea-app/ruletemplates/timeeventitemtemplate.h
@@ -64,7 +64,7 @@ class TimeEventItemTemplates: public QAbstractListModel
 
 public:
     TimeEventItemTemplates(QObject *parent = nullptr): QAbstractListModel(parent) {}
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent) return m_list.count(); }
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent) return static_cast<int>(m_list.count()); }
     QVariant data(const QModelIndex &index, int role) const override { Q_UNUSED(index) Q_UNUSED(role) return QVariant(); }
 
     Q_INVOKABLE TimeEventItemTemplate* get(int index) const {
@@ -76,7 +76,7 @@ public:
 
     void addTimeEventItemTemplate(TimeEventItemTemplate *timeEventItemTemplate) {
         timeEventItemTemplate->setParent(this);
-        beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+        beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
         m_list.append(timeEventItemTemplate);
         endInsertRows();
     }
