+++ libnymea-app/types/eventdescriptors.cpp
@@ -36,7 +36,7 @@ EventDescriptors::EventDescriptors(QObject *parent) :
 int EventDescriptors::rowCount(const QModelIndex &parent) const
 {
     Q_UNUSED(parent)
-    return m_list.count();
+    return static_cast<int>(m_list.count());
 }
 
 QVariant EventDescriptors::data(const QModelIndex &index, int role) const
@@ -74,7 +74,7 @@ EventDescriptor *EventDescriptors::createNewEventDescriptor()
 void EventDescriptors::addEventDescriptor(EventDescriptor *eventDescriptor)
 {
     eventDescriptor->setParent(this);
-    beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+    beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
     m_list.append(eventDescriptor);
     endInsertRows();
     emit countChanged();
